pyproject.toml
[project]
name = "openguard"
version = "0.1.1"
description = "guarding proxy for AI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "OpenGuard Contributors" },
]
keywords = ["llm", "proxy", "guardrails", "openai", "anthropic", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: Proxy Servers",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.6",
    "httpx==0.27.0",
    "cachetools==5.5.2",
    "asyncache==0.3.1",
    "pydantic>=2.11.0",
    "pyyaml==6.0.2",
    "typer>=0.12.0",
]

[project.urls]
Homepage = "https://github.com/Jitera-Labs/openguard"
Repository = "https://github.com/Jitera-Labs/openguard"
Issues = "https://github.com/Jitera-Labs/openguard/issues"

[project.optional-dependencies]
dev = [
    "pytest==8.3.0",
    "pytest-asyncio==0.24.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "livereload>=2.7.1",
    "types-cachetools>=6.2.0.20251022",
    "types-pyyaml>=6.0.12.20250915",
]

[project.scripts]
openguard = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/pyproject.toml",
    "/guards.yaml.example",
    "/presets",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
# E: pycodestyle, F: pyflakes, I: isort
select = ["E", "F", "I"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
check_untyped_defs = false
warn_return_any = false
warn_unused_ignores = false
show_error_codes = true

[[tool.mypy.overrides]]
module = "src.dotty"
disable_error_code = ["arg-type", "assignment"]

[dependency-groups]
dev = [
    "types-cachetools>=6.2.0.20251022",
    "types-pyyaml>=6.0.12.20250915",
]
